+++ src/server/handlers/mod.rs
@@ -3,6 +3,7 @@ pub mod config;
 pub mod exitnode;
 pub mod local;
 pub mod local_config;
+pub mod logs;
 pub mod metrics;
 pub mod physnet;
 pub mod system;